+++ src/item.cpp
@@ -194,7 +194,7 @@ item::item() : bday( calendar::start_of_cataclysm )
 {
     type = nullitem();
     charges = 0;
-    contents = item_contents( type->pocket_data );
+    contents = item_contents( type->pockets );
 }
 
 item::item( const itype *type, time_point turn, int qty ) : type( type ), bday( turn )
@@ -260,7 +260,7 @@ item::item( const itype *type, time_point turn, int qty ) : type( type ), bday( 
     if( type->relic_data ) {
         relic_data = *type->relic_data;
     }
-    contents = type->pocket_data;
+    contents = item_contents( type->pockets );
 }
 
 item::item( const itype_id &id, time_point turn, int qty )
