+++ src/item_factory.cpp
@@ -17,7 +17,6 @@
 #include "catacharset.h"
 #include "debug.h"
 #include "enums.h"
-#include "generic_factory.h"
 #include "init.h"
 #include "item.h"
 #include "item_category.h"
@@ -2189,7 +2188,7 @@ void Item_factory::load_basic_info( const JsonObject &jo, itype &def, const std:
         }
     }
 
-    assign( jo, "pocket_data", def.pocket_data );
+    assign( jo, "pocket_data", def.pockets );
 
     load_slot_optional( def.container, jo, "container_data", src );
     load_slot_optional( def.armor, jo, "armor_data", src );
